ER-1305 short description (#637)
* ER-1304-New screen.
