+++ src/Employer/Employer.Web/Controllers/Part1/NumberOfPositionsController.cs
@@ -45,7 +45,7 @@ namespace Esfa.Recruit.Employer.Web.Controllers.Part1
             }
 
             return wizard
-                ? RedirectToRoute(RouteNames.ShortDescription_Get, new {vacancyId = response.Data})
+                ? RedirectToRoute(RouteNames.Employer_Get, new {vacancyId = response.Data})
                 : RedirectToRoute(RouteNames.Vacancy_Preview_Get);
         }
     }
