+++ src/Employer/Employer.Web/Orchestrators/Part2/ShortDescriptionOrchestrator.cs
@@ -1,0 +1,80 @@
+using System.Threading.Tasks;
+using Esfa.Recruit.Employer.Web.Configuration.Routing;
+using Esfa.Recruit.Employer.Web.Mappings;
+using Esfa.Recruit.Employer.Web.RouteModel;
+using Esfa.Recruit.Employer.Web.ViewModels.Part2.ShortDescription;
+using Esfa.Recruit.Shared.Web.Orchestrators;
+using Esfa.Recruit.Shared.Web.Services;
+using Esfa.Recruit.Vacancies.Client.Application.Validation;
+using Esfa.Recruit.Vacancies.Client.Domain.Entities;
+using Esfa.Recruit.Vacancies.Client.Infrastructure.Client;
+using Microsoft.Extensions.Logging;
+
+namespace Esfa.Recruit.Employer.Web.Orchestrators.Part2
+{
+    public class ShortDescriptionOrchestrator : EntityValidatingOrchestrator<Vacancy, ShortDescriptionEditModel>
+    {
+        private const VacancyRuleSet ValidationRules = VacancyRuleSet.ShortDescription;
+        private readonly IEmployerVacancyClient _client;
+        private readonly IRecruitVacancyClient _vacancyClient;
+        private readonly IReviewSummaryService _reviewSummaryService;
+
+        public ShortDescriptionOrchestrator(IEmployerVacancyClient client, IRecruitVacancyClient vacancyClient, ILogger<ShortDescriptionOrchestrator> logger, IReviewSummaryService reviewSummaryService) : base(logger)
+        {
+            _client = client;
+            _vacancyClient = vacancyClient;
+            _reviewSummaryService = reviewSummaryService;
+        }
+
+        public async Task<ShortDescriptionViewModel> GetShortDescriptionViewModelAsync(VacancyRouteModel vrm)
+        {
+            var vacancy = await Utility.GetAuthorisedVacancyForEditAsync(_client, _vacancyClient, vrm, RouteNames.ShortDescription_Get);
+
+            var vm = new ShortDescriptionViewModel
+            {
+                VacancyId = vacancy.Id,
+                ShortDescription = vacancy.ShortDescription,
+                Title = vacancy.Title
+            };
+
+            if (vacancy.Status == VacancyStatus.Referred)
+            {
+                vm.Review = await _reviewSummaryService.GetReviewSummaryViewModelAsync(vacancy.VacancyReference.Value, 
+                    ReviewFieldMappingLookups.GetShortDescriptionReviewFieldIndicators());
+            }
+
+            return vm;
+        }
+
+        public async Task<ShortDescriptionViewModel> GetShortDescriptionViewModelAsync(ShortDescriptionEditModel m)
+        {
+            var vm = await GetShortDescriptionViewModelAsync((VacancyRouteModel)m);
+            
+            vm.ShortDescription = m.ShortDescription;
+
+            return vm;
+        }
+
+        public async Task<OrchestratorResponse> PostShortDescriptionEditModelAsync(ShortDescriptionEditModel m, VacancyUser user)
+        {
+            var vacancy = await Utility.GetAuthorisedVacancyForEditAsync(_client, _vacancyClient, m, RouteNames.ShortDescription_Post);
+
+            vacancy.ShortDescription = m.ShortDescription;
+
+            return await ValidateAndExecute(
+                vacancy, 
+                v => _vacancyClient.Validate(v, ValidationRules),
+                v => _vacancyClient.UpdateDraftVacancyAsync(vacancy, user)
+            );
+        }
+
+        protected override EntityToViewModelPropertyMappings<Vacancy, ShortDescriptionEditModel> DefineMappings()
+        {
+            var mappings = new EntityToViewModelPropertyMappings<Vacancy, ShortDescriptionEditModel>();
+
+            mappings.Add(e => e.ShortDescription, vm => vm.ShortDescription);
+
+            return mappings;
+        }
+    }
+}
