+++ src/Employer/Employer.Web/ViewModels/DisplayVacancyViewModel.cs
