--- src/Employer/Employer.Web/ViewModels/Part1/ShortDescription/ShortDescriptionEditModel.cs
@@ -1,9 +1,0 @@
-﻿using Esfa.Recruit.Employer.Web.RouteModel;
-
-namespace Esfa.Recruit.Employer.Web.ViewModels.Part1.ShortDescription
-{
-    public class ShortDescriptionEditModel : VacancyRouteModel
-    {
-        public string ShortDescription { get; set; }
-    }
-}
