--- src/Employer/Employer.Web/Views/Part1/ShortDescription/ShortDescription.cshtml
@@ -1,44 +1,0 @@
-﻿@using Esfa.Recruit.Shared.Web.Mappers
-@model Esfa.Recruit.Employer.Web.ViewModels.Part1.ShortDescription.ShortDescriptionViewModel
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        
-        <a asp-show="@Model.PageInfo.IsWizard" asp-route="@RouteNames.NumberOfPositions_Get" esfa-automation="link-back" class="govuk-back-link govuk-!-margin-top-0 govuk-!-margin-bottom-6">Back</a>
-
-        <partial name="@PartialNames.ValidationSummary" model='new ValidationSummaryViewModel { ModelState = ViewData.ModelState, OrderedFieldNames = Model.OrderedFieldNames }'/>
-
-        <partial name="@PartialNames.ReviewSummary" for="Review"/>
-
-        <h1 class="govuk-heading-xl govuk-!-margin-bottom-6">Brief overview</h1>
-    </div>
-</div>
-
-<div class="govuk-grid-row">
-    <div class="govuk-grid-column-two-thirds">
-        <form asp-route="@RouteNames.ShortDescription_Post" asp-route-wizard="@Model.PageInfo.IsWizard" novalidate>
-             <div class="govuk-character-count"> 
-                <div esfa-validation-marker-for="ShortDescription" class="govuk-form-group">
-                    <label asp-for="ShortDescription" class="govuk-label govuk-!-font-weight-bold">
-                        Write a few sentences to describe your vacancy
-                    </label>
-                    <span class="govuk-hint">Brief overview needs to be engaging to make the candidate want to view the full vacancy.</span>
-                    <span esfa-validation-message-for="ShortDescription" class="govuk-error-message"></span>
-                    <textarea asp-for="ShortDescription" class="govuk-textarea character-count" rows="5" data-val-length-max="350" esfa-automation="short-description"></textarea>
-                    
-                    <div>
-                        <div class="govuk-body float-left">Minimum 50 characters</div>
-                        <div class="govuk-body float-right">
-                            <span class="maxchar-count">350</span>
-                            <span class="maxchar-text"> characters remaining</span>
-                        </div>
-                    </div>
-                    <div class="clear"></div>
-                </div>
-            </div> 
-            <button type="submit" class="govuk-button save-button" esfa-automation="continue-button">@Model.PageInfo.SubmitButtonText</button>
-            <a asp-show="@Model.PageInfo.IsWizard" asp-route="@RouteNames.Dashboard_Index_Get" class="govuk-link das-button-link" esfa-automation="cancel-link">Cancel</a>
-            <a asp-show="@Model.PageInfo.IsNotWizard" asp-route="@RouteNames.Vacancy_Preview_Get" asp-fragment="@Anchors.ShortDescriptionSection" class="govuk-link das-button-link" esfa-automation="cancel-link">Cancel</a>
-        </form>
-    </div>
-</div>
