+++ src/Employer/UnitTests/Employer.Web/UtilityTests/CheckRouteIsValidForVacancyTests.cs
@@ -77,29 +77,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
         [InlineData(RouteNames.Title_Post, false)]
         [InlineData(RouteNames.NumberOfPositions_Get, false)]
         [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
-        [InlineData("any other route", true)]
-        public void ShouldRedirectToShortDescription(string route, bool shouldRedirect)
-        {
-            var vacancy = new Vacancy
-            {
-                EmployerAccountId = "EMPLOYER ACCOUNT ID",
-                Id = Guid.Parse("84af954e-5baf-4942-897d-d00180a0839e"),
-                Title = "has a value",
-                NumberOfPositions = 3
-            };
-
-            CheckRouteIsValidForVacancyTest(vacancy, route, shouldRedirect, RouteNames.ShortDescription_Get);
-        }
-
-        [Theory]
-        [InlineData(RouteNames.Title_Get, false)]
-        [InlineData(RouteNames.Title_Post, false)]
-        [InlineData(RouteNames.NumberOfPositions_Get, false)]
-        [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
         [InlineData(RouteNames.Employer_Get, false)]
         [InlineData(RouteNames.Employer_Post, false)]
         [InlineData("any other route", true)]
@@ -110,8 +87,7 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 EmployerAccountId = "EMPLOYER ACCOUNT ID",
                 Id = Guid.Parse("84af954e-5baf-4942-897d-d00180a0839e"),
                 Title = "has a value",
-                NumberOfPositions = 3,
-                ShortDescription = "has a value"
+                NumberOfPositions = 3
             };
 
             CheckRouteIsValidForVacancyTest(vacancy, route, shouldRedirect, RouteNames.Employer_Get);
@@ -122,8 +98,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
         [InlineData(RouteNames.Title_Post, false)]
         [InlineData(RouteNames.NumberOfPositions_Get, false)]
         [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
         [InlineData(RouteNames.Employer_Get, false)]
         [InlineData(RouteNames.Employer_Post, false)]
         [InlineData(RouteNames.EmployerName_Get, false)]
@@ -144,8 +118,7 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 NumberOfPositions = 3,
                 LegalEntityName = "Legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
-                EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value"
+                EmployerLocation = new Address { Postcode = "has a value" }
             };
 
             CheckRouteIsValidForVacancyTest(vacancy, route, shouldRedirect, RouteNames.Training_Get);
@@ -156,8 +129,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
         [InlineData(RouteNames.Title_Post, false)]
         [InlineData(RouteNames.NumberOfPositions_Get, false)]
         [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
         [InlineData(RouteNames.Employer_Get, false)]
         [InlineData(RouteNames.Employer_Post, false)]
         [InlineData(RouteNames.EmployerName_Get, false)]
@@ -181,7 +152,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 LegalEntityName = "legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 ProgrammeId = "has a value"
             };
 
@@ -201,7 +171,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 LegalEntityName = "legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 ProgrammeId = "has a value",
                 Wage = new Wage { WageType = WageType.FixedWage}
             };
@@ -220,7 +189,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 NumberOfPositions = 3,
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 ProgrammeId = "has a value",
                 Wage = new Wage { WageType = WageType.FixedWage}
             };
@@ -239,7 +207,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 NumberOfPositions = 3,
                 LegalEntityName = "legal name",
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 ProgrammeId = "has a value",
                 Wage = new Wage { WageType = WageType.FixedWage}
             };
@@ -258,7 +225,6 @@ namespace Esfa.Recruit.Employer.UnitTests.Employer.Web.UtilityTests
                 NumberOfPositions = 3,
                 LegalEntityName = "legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
-                ShortDescription = "has a value",
                 ProgrammeId = "has a value",
                 Wage = new Wage { WageType = WageType.FixedWage}
             };
