+++ src/Employer/UnitTests/Employer.Web/ViewModels/TrainingProvider/ConfirmTrainingProviderEditModelTests.cs
@@ -1,9 +1,5 @@
-﻿using Esfa.Recruit.Employer.Web.ViewModels;
-using FluentAssertions;
+﻿using FluentAssertions;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
 using Esfa.Recruit.Employer.Web.ViewModels.Part1.TrainingProvider;
 using Xunit;
