+++ src/Provider/Provider.Web/Configuration/ProviderApprenticeshipsLinkHelper.cs
@@ -13,7 +13,6 @@ namespace Esfa.Recruit.Provider.Web.Configuration
             _externalLinks = externalLinks.Value;
             _pasRoutes = pasRoutes.Value;
         }
-
         public string AccountHome => $"{_externalLinks.ProviderApprenticeshipSiteUrl}{_pasRoutes.ProviderApprenticeshipSiteAccountsHomeRoute}";
         public string Notifications => $"{_externalLinks.ProviderApprenticeshipSiteUrl}{_pasRoutes.ProviderApprenticeshipSiteNotificationSettingsRoute}";
         public string Apprentices => $"{_externalLinks.CommitmentsSiteUrl}{_pasRoutes.ProviderApprenticeshipSiteManageApprenticesRoute}";
