+++ src/Provider/Provider.Web/Controllers/Part1/NumberOfPositionsController.cs
@@ -48,7 +48,7 @@ namespace Esfa.Recruit.Provider.Web.Controllers.Part1
             }
 
             return wizard
-                ? RedirectToRoute(RouteNames.ShortDescription_Get, new {vacancyId = response.Data})
+                ? RedirectToRoute(RouteNames.LegalEntity_Get, new {vacancyId = response.Data})
                 : RedirectToRoute(RouteNames.Vacancy_Preview_Get);
         }    
     }
