+++ src/Provider/Provider.Web/Controllers/Part2/ShortDescriptionController.cs
@@ -3,12 +3,11 @@ using Esfa.Recruit.Provider.Web.Configuration.Routing;
 using Esfa.Recruit.Provider.Web.Extensions;
 using Esfa.Recruit.Provider.Web.Orchestrators.Part1;
 using Esfa.Recruit.Provider.Web.RouteModel;
-using Esfa.Recruit.Provider.Web.ViewModels.Part1.ShortDescription;
+using Esfa.Recruit.Provider.Web.ViewModels.Part2.ShortDescription;
 using Esfa.Recruit.Shared.Web.Extensions;
-using Esfa.Recruit.Shared.Web.Mappers;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Esfa.Recruit.Provider.Web.Controllers.Part1
+namespace Esfa.Recruit.Provider.Web.Controllers.Part2
 {
     [Route(RoutePaths.AccountVacancyRoutePath)]
     public class ShortDescriptionController : Controller
@@ -21,15 +20,14 @@ namespace Esfa.Recruit.Provider.Web.Controllers.Part1
         }
         
         [HttpGet("description", Name = RouteNames.ShortDescription_Get)]
-        public async Task<IActionResult> ShortDescription(VacancyRouteModel vrm, [FromQuery] string wizard = "true")
+        public async Task<IActionResult> ShortDescription(VacancyRouteModel vrm)
         {
             var vm = await _orchestrator.GetShortDescriptionViewModelAsync(vrm);
-            vm.PageInfo.SetWizard(wizard);
             return View(vm);
         }
 
         [HttpPost("description", Name = RouteNames.ShortDescription_Post)]
-        public async Task<IActionResult> ShortDescription(ShortDescriptionEditModel m, [FromQuery] bool wizard)
+        public async Task<IActionResult> ShortDescription(ShortDescriptionEditModel m)
         {
             var response = await _orchestrator.PostShortDescriptionEditModelAsync(m, User.ToVacancyUser());
 
@@ -41,13 +39,10 @@ namespace Esfa.Recruit.Provider.Web.Controllers.Part1
             if (!ModelState.IsValid)
             {
                 var vm = await _orchestrator.GetShortDescriptionViewModelAsync(m);
-                vm.PageInfo.SetWizard(wizard);
                 return View(vm);
             }
 
-            return wizard 
-                ? RedirectToRoute(RouteNames.LegalEntity_Get)
-                : RedirectToRoute(RouteNames.Vacancy_Preview_Get, null, Anchors.ShortDescriptionSection);
+            return RedirectToRoute(RouteNames.Vacancy_Preview_Get);
         }
     }
 }
