+++ src/Provider/Provider.Web/ViewModels/Part1/Title/TitleViewModel.cs
@@ -1,9 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using Esfa.Recruit.Provider.Web.Configuration.Routing;
 using Esfa.Recruit.Shared.Web.ViewModels;
-using Esfa.Recruit.Vacancies.Client.Infrastructure.QueryStore.Projections;
 
 namespace Esfa.Recruit.Provider.Web.ViewModels.Part1.Title
 {
@@ -19,9 +17,9 @@ namespace Esfa.Recruit.Provider.Web.ViewModels.Part1.Title
         };
         public PartOnePageInfoViewModel PageInfo { get; set; }
         public string FormPostRouteName => VacancyId.HasValue ? RouteNames.Title_Post : RouteNames.CreateVacancy_Post;
-        public bool HasCloneableVacancies { get; set; }
+        public bool HasAnyVacancies { get; set; }
         public ReviewSummaryViewModel Review { get; set; } = new ReviewSummaryViewModel();
         public string BackLink =>
-            HasCloneableVacancies ? RouteNames.Vacancies_Get : RouteNames.CreateVacancy_Get;
+            HasAnyVacancies ? RouteNames.Vacancies_Get : RouteNames.CreateVacancy_Get;
     }
 }
