+++ src/Provider/Provider.Web/ViewModels/Part2/ShortDescription/ShortDescriptionEditModel.cs
@@ -1,0 +1,9 @@
+using Esfa.Recruit.Provider.Web.RouteModel;
+
+namespace Esfa.Recruit.Provider.Web.ViewModels.Part2.ShortDescription
+{
+    public class ShortDescriptionEditModel : VacancyRouteModel
+    {
+        public string ShortDescription { get; set; }
+    }
+}
