+++ src/Provider/UnitTests/Provider.Web/UtilityTests/CheckRouteIsValidForVacancyTests.cs
@@ -48,29 +48,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
         [InlineData(RouteNames.Title_Post, false)]
         [InlineData(RouteNames.NumberOfPositions_Get, false)]
         [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
-        [InlineData("any other route", true)]
-        public void ShouldRedirectToShortDescription(string route, bool shouldRedirect)
-        {
-            var vacancy = new Vacancy 
-            {
-                EmployerAccountId = "EMPLOYER ACCOUNT ID",
-                Id = Guid.Parse("84af954e-5baf-4942-897d-d00180a0839e"),
-                Title = "has a value",
-                NumberOfPositions = 1
-            };
-
-            CheckRouteIsValidForVacancyTest(vacancy, route, shouldRedirect, RouteNames.ShortDescription_Get);
-        }
-
-        [Theory]
-        [InlineData(RouteNames.Title_Get, false)]
-        [InlineData(RouteNames.Title_Post, false)]
-        [InlineData(RouteNames.NumberOfPositions_Get, false)]
-        [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
         [InlineData(RouteNames.Location_Get, false)]
         [InlineData(RouteNames.Location_Post, false)]
         [InlineData("any other route", true)]
@@ -81,8 +58,7 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
                 EmployerAccountId = "EMPLOYER ACCOUNT ID",
                 Id = Guid.Parse("84af954e-5baf-4942-897d-d00180a0839e"),
                 Title = "has a value",
-                NumberOfPositions = 1,
-                ShortDescription = "has a value"
+                NumberOfPositions = 1
             };
 
             CheckRouteIsValidForVacancyTest(vacancy, route, shouldRedirect, RouteNames.LegalEntity_Get);
@@ -93,8 +69,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
         [InlineData(RouteNames.Title_Post, false)]
         [InlineData(RouteNames.NumberOfPositions_Get, false)]
         [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
         [InlineData(RouteNames.LegalEntity_Get, false)]
         [InlineData(RouteNames.LegalEntity_Post, false)]
         [InlineData(RouteNames.EmployerName_Get, false)]
@@ -112,7 +86,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
                 Id = Guid.Parse("84af954e-5baf-4942-897d-d00180a0839e"),
                 Title = "has a value",
                 NumberOfPositions = 1,
-                ShortDescription = "has a value",
                 LegalEntityName = "legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 EmployerLocation = new Address { Postcode = "CV1 2WT" }
@@ -126,8 +99,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
         [InlineData(RouteNames.Title_Post, false)]
         [InlineData(RouteNames.NumberOfPositions_Get, false)]
         [InlineData(RouteNames.NumberOfPositions_Post, false)]
-        [InlineData(RouteNames.ShortDescription_Get, false)]
-        [InlineData(RouteNames.ShortDescription_Post, false)]
         [InlineData(RouteNames.LegalEntity_Get, false)]
         [InlineData(RouteNames.LegalEntity_Post, false)]
         [InlineData(RouteNames.EmployerName_Get, false)]
@@ -148,7 +119,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
                 Title = "has a value",
                 NumberOfPositions = 1,
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 LegalEntityName = "legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 ProgrammeId = "has a value"
@@ -168,7 +138,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
                 Title = "has a value",
                 NumberOfPositions = 1,
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 LegalEntityName = "legal name",
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 ProgrammeId = "has a value",
@@ -189,7 +158,6 @@ namespace Esfa.Recruit.Provider.UnitTests.Provider.Web.UtilityTests
                 NumberOfPositions = 1,
                 EmployerNameOption = EmployerNameOption.RegisteredName,
                 EmployerLocation = new Address { Postcode = "has a value" },
-                ShortDescription = "has a value",
                 ProgrammeId = "has a value",
                 Wage = new Wage { WageType = WageType.FixedWage}
             };
